+++ packages/rest-generator/src/parser/refs.spec.ts
@@ -1,18 +1,5 @@
-import { resolve, $Refs } from '@apidevtools/swagger-parser';
-import { OpenAPIV3 } from 'openapi-types';
+import { createRefs } from '../../test/test-util';
 import { parseTypeName, resolveObject } from './refs';
-
-export const emptyApiDefinition = {
-  openapi: '3.0.0',
-  info: { title: 'Spec', version: '1.0.0' },
-  paths: {}
-} as const;
-
-export function createRefs(
-  components: OpenAPIV3.ComponentsObject = {}
-): Promise<$Refs> {
-  return resolve({ ...emptyApiDefinition, components });
-}
 
 describe('parseTypeName', () => {
   it('get the last part of a reference as type name', () => {
